Add files via upload
The Company observed that the employees are leaving eventually so they provided data and asked to create a report for the attrition of employees. So, what would be the best way than to create a dashboard to show the complexity of problem in an easy way. In this Dashboard we have used various Visualization charts to find out the Attrition rate in a company by Age, Gender, Department, Education and salary.